requirements.txt
apache_beam==2.25.0
brotli==1.0.9
cryptography==3.2.1
Cython==0.29.21
docutils==0.16
Flask==1.1.2
gunicorn==20.0.4
importlib_metadata==2.0.0
ipaddr==2.2.0
ipython==7.19.0
keyring==21.5.0
lockfile==0.12.2
lxml==4.6.1
matplotlib==3.3.3
mock==4.0.2
mypy_extensions==0.4.3
nets==0.0.3
numpy==1.19.4
ordereddict==1.1
pandas==1.1.4
Pillow==8.0.1
protobuf==3.13.0
pycocotools==2.0.2
pyOpenSSL==19.1.0
scipy==1.5.4
simplejson==3.17.2
tensorflow==2.3.1
tf_slim==1.1.0
typing_extensions==3.7.4.3
wincertstore==0.2
zipp==3.4.0
